Module 3 Activity Complete
Can't figure out why success function isn't running to display the
geojson data. I thought it was was maybe something to do with how the
file was accessed, so I moved it in js folder so file path can't be
messed up. I'm not sure what else to do as it seems the code is fine.
Who knows...