Module 5b Complete (5th operator)
I chose to go for the compare operator, using a combination of lists
and bar graphs to generate the comparison. The two js files;
chartData.js and index.js contain the code for the comparison operator.
Beware though… the code is not optimized, and is probably hard to read.
Didn’t have time to comment and will comment and organize everything
this weekend when I finish my styling and content.